@@ -19,16 +19,16 @@
     "@agm/core": "^3.0.0-beta.0",
     "@agm/snazzy-info-window": "^3.0.0-beta.0",
     "@angular-eslint/schematics": "^0.8.0-beta.3",
-    "@angular/animations": "^10.2.3",
+    "@angular/animations": "^11.0.3",
     "@angular/cdk": "^10.2.7",
-    "@angular/common": "^10.2.3",
-    "@angular/compiler": "^10.2.3",
-    "@angular/core": "^10.2.3",
-    "@angular/forms": "^10.2.3",
+    "@angular/common": "^11.0.3",
+    "@angular/compiler": "^11.0.3",
+    "@angular/core": "^11.0.3",
+    "@angular/forms": "^11.0.3",
     "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "^10.2.3",
-    "@angular/platform-browser-dynamic": "^10.2.3",
-    "@angular/router": "^10.2.3",
+    "@angular/platform-browser": "^11.0.3",
+    "@angular/platform-browser-dynamic": "^11.0.3",
+    "@angular/router": "^11.0.3",
     "@apollo/client": "^3.2.7",
     "@clr/angular": "^4.0.6",
     "@clr/core": "^3.1.12",
@@ -67,24 +67,24 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^8.4.1",
-    "@angular-devkit/build-angular": "~0.1002.0",
-    "@angular/cli": "~10.2.0",
-    "@angular/compiler-cli": "^10.2.3",
-    "@angular/language-service": "^10.2.3",
+    "@angular-devkit/build-angular": "~0.1100.3",
+    "@angular/cli": "~11.0.3",
+    "@angular/compiler-cli": "^11.0.3",
+    "@angular/language-service": "^11.0.3",
     "@graphql-codegen/cli": "^1.19.2",
     "@graphql-codegen/fragment-matcher": "2.0.0",
     "@graphql-codegen/typescript": "^1.17.11",
     "@graphql-codegen/typescript-apollo-angular": "^2.1.0",
     "@graphql-codegen/typescript-operations": "^1.17.10",
-    "@types/jasmine": "~3.3.8",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.19.5",
     "graphql-cli": "^4.0.0-alpha",
     "graphql-toolkit": "^0.7.5",
     "html-webpack-plugin": "^3.2.0",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~5.1.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
