@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ApolloTestingController, ApolloTestingModule } from 'apollo-angular/testing';
 
 import { BlogpostComponent } from './blogpost.component';
@@ -12,10 +12,10 @@ describe('BlogpostComponent', () => {
   let fixture: ComponentFixture<BlogpostComponent>;
   let backend: ApolloTestingController;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ BlogpostComponent ],
-      imports: [ MomentModule, MomentTimezoneModule, RouterModule.forRoot([]), ApolloTestingModule, HttpClientModule ]
+      imports: [ MomentModule, MomentTimezoneModule, RouterModule.forRoot([], { relativeLinkResolution: 'legacy' }), ApolloTestingModule, HttpClientModule ]
     })
     .compileComponents();
   }));
