@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ApolloTestingController, ApolloTestingModule } from 'apollo-angular/testing';
 import { PolicyComponent } from './policy.component';
 import { RouterModule } from '@angular/router';
@@ -9,10 +9,10 @@ describe('PolicyComponent', () => {
   let fixture: ComponentFixture<PolicyComponent>;
   let backend: ApolloTestingController;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PolicyComponent ],
-      imports: [ RouterModule.forRoot([]), ApolloTestingModule, HttpClientModule ]
+      imports: [ RouterModule.forRoot([], { relativeLinkResolution: 'legacy' }), ApolloTestingModule, HttpClientModule ]
     })
     .compileComponents();
   }));
