@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ApolloTestingController, ApolloTestingModule } from 'apollo-angular/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 
@@ -12,7 +12,7 @@ describe('BlogsComponent', () => {
   let fixture: ComponentFixture<BlogsComponent>;
   let backend: ApolloTestingController;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ BlogsComponent ],
       imports: [ MomentModule, MomentTimezoneModule, ApolloTestingModule, RouterTestingModule ]
