@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ApolloTestingController, ApolloTestingModule } from 'apollo-angular/testing';
 
 import { BoardmembersComponent } from './boardmembers.component';
@@ -8,7 +8,7 @@ describe('BoardmembersComponent', () => {
   let fixture: ComponentFixture<BoardmembersComponent>;
   let backend: ApolloTestingController;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ BoardmembersComponent ],
       imports: [ ApolloTestingModule ]
