@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ConventionsComponent } from './conventions.component';
 import { MatTableModule } from '@angular/material/table';
@@ -15,7 +15,7 @@ describe('ConventionsComponent', () => {
   let fixture: ComponentFixture<ConventionsComponent>;
   let backend: ApolloTestingController;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ConventionsComponent ],
       imports: [ 
