@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { GeekwaymicroComponent } from './geekwaymicro.component';
 import { MomentModule } from 'ngx-moment';
@@ -17,7 +17,7 @@ describe('GeekwaymicroComponent', () => {
   let fixture: ComponentFixture<GeekwaymicroComponent>;
   let backend: ApolloTestingController;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ GeekwaymicroComponent ],
       imports: [ 
