@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ApolloTestingController, ApolloTestingModule } from 'apollo-angular/testing';
 
 import { GeekwayminiComponent } from './geekwaymini.component';
@@ -17,7 +17,7 @@ describe('GeekwayminiComponent', () => {
   let fixture: ComponentFixture<GeekwayminiComponent>;
   let backend: ApolloTestingController;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ GeekwayminiComponent ],
       imports: [ 
