@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ApolloTestingController, ApolloTestingModule } from 'apollo-angular/testing';
 import { MatCardModule } from '@angular/material/card';
 
@@ -9,7 +9,7 @@ describe('EventsComponent', () => {
   let fixture: ComponentFixture<EventsComponent>;
   let backend: ApolloTestingController;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ EventsComponent ],
       imports: [ ApolloTestingModule, MatCardModule ]
