@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { RedirectionMessageComponent } from './redirection-message.component';
 
@@ -6,7 +6,7 @@ describe('RedirectionMessageComponent', () => {
   let component: RedirectionMessageComponent;
   let fixture: ComponentFixture<RedirectionMessageComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ RedirectionMessageComponent ]
     })
